+++ Libplanet/Blockchain/BlockChain.MineBlock.cs
@@ -6,14 +6,12 @@ using System.Diagnostics;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Bencodex;
 using Bencodex.Types;
 using Libplanet.Action;
 using Libplanet.Blockchain.Policies;
 using Libplanet.Blocks;
 using Libplanet.Crypto;
 using Libplanet.Tx;
-using static Libplanet.Blocks.BlockMarshaler;
 
 namespace Libplanet.Blockchain
 {
@@ -133,18 +131,7 @@ namespace Libplanet.Blockchain
                 difficulty,
                 prevHash);
 
-            var metadata = new BlockMetadata
-            {
-                Index = index,
-                Difficulty = difficulty,
-                TotalDifficulty = Tip.TotalDifficulty + difficulty,
-                PublicKey = miner.PublicKey,
-                PreviousHash = prevHash,
-                Timestamp = timestamp,
-            };
-
             var transactionsToMine = GatherTransactionsToMine(
-                metadata,
                 maxBlockBytes: maxBlockBytes,
                 maxTransactions: maxTransactions,
                 maxTransactionsPerSigner: maxTransactionsPerSigner,
@@ -168,7 +155,16 @@ namespace Libplanet.Blockchain
                 index,
                 transactionsToMine.Count);
 
-            var blockContent = new BlockContent<T>(metadata) { Transactions = transactionsToMine };
+            var blockContent = new BlockContent<T>
+            {
+                Index = index,
+                Difficulty = difficulty,
+                TotalDifficulty = Tip.TotalDifficulty + difficulty,
+                PublicKey = miner.PublicKey,
+                PreviousHash = prevHash,
+                Timestamp = timestamp,
+                Transactions = transactionsToMine,
+            };
             PreEvaluationBlock<T> preEval;
 
             TipChanged += WatchTip;
@@ -227,7 +223,6 @@ namespace Libplanet.Blockchain
         /// Gathers <see cref="Transaction{T}"/>s for mining a next block
         /// from the current set of staged <see cref="Transaction{T}"/>s.
         /// </summary>
-        /// <param name="metadata">The metadata of the block to be mined.</param>
         /// <param name="maxBlockBytes">The maximum number of bytes a block can have.</param>
         /// <param name="maxTransactions">The maximum number of <see cref="Transaction{T}"/>s
         /// allowed.</param>
@@ -240,7 +235,6 @@ namespace Libplanet.Blockchain
         /// <see cref="Transaction{T}"/>s in the list for each signer not exceeding
         /// <paramref name="maxTransactionsPerSigner"/>.</returns>
         internal ImmutableList<Transaction<T>> GatherTransactionsToMine(
-            BlockMetadata metadata,
             long maxBlockBytes,
             int maxTransactions,
             int maxTransactionsPerSigner,
@@ -260,30 +254,8 @@ namespace Libplanet.Blockchain
             // FIXME: The tx collection timeout should be configurable.
             DateTimeOffset timeout = DateTimeOffset.UtcNow + TimeSpan.FromSeconds(4);
 
-            int digestSize = BlockMetadata.HashAlgorithmType.DigestSize;
-
-            // Makes an empty block payload to estimate the length of bytes without transactions.
-            // FIXME: We'd better to estimate only transactions rather than the whole block.
-            var dumbSig = metadata.PublicKey is null
-                ? (ImmutableArray<byte>?)null
-                : ImmutableArray.Create(new byte[71]);
-            Bencodex.Types.Dictionary marshaledEmptyBlock = MarshalBlock(
-                marshaledBlockHeader: MarshalBlockHeader(
-                    marshaledPreEvaluatedBlockHeader: MarshalPreEvaluationBlockHeader(
-                        marshaledMetadata: MarshalBlockMetadata(metadata),
-                        nonce: default,
-                        preEvaluationHash: new byte[digestSize].ToImmutableArray()
-                    ),
-                    stateRootHash: default,
-                    signature: dumbSig,
-                    hash: default
-                ),
-                marshaledTransactions: BlockMarshaler.MarshalTransactions(
-                    Array.Empty<Transaction<T>>()
-                )
-            );
-            Dictionary estimatedEncoding = marshaledEmptyBlock;
-
+            List estimatedEncoding = BlockMarshaler.MarshalTransactions(
+                new List<Transaction<T>>());
             var storedNonces = new Dictionary<Address, long>();
             var nextNonces = new Dictionary<Address, long>();
             var toMineCounts = new Dictionary<Address, int>();
@@ -333,22 +305,21 @@ namespace Libplanet.Blockchain
                         continue;
                     }
 
-                    Dictionary txAddedBlockEncoding =
-                        AppendTxToMarshaledBlock(estimatedEncoding, tx);
-                    if (txAddedBlockEncoding.EncodingLength > maxBlockBytes)
+                    List txAddedEncoding = estimatedEncoding.Add(
+                        BlockMarshaler.MarshalTransaction(tx));
+                    if (txAddedEncoding.EncodingLength > maxBlockBytes)
                     {
                         _logger.Debug(
-                            "Ignoring tx {Iter}/{Total} {TxId} due to the maximum size " +
-                            "allowed for a block: {CurrentEstimate}/{MaximumBlockBytes}",
+                            "Ignoring tx {Iter}/{Total} {TxId} due to the maximum size allowed " +
+                            "for transactions in a block: {CurrentEstimate}/{MaximumBlockBytes}",
                             i,
                             stagedTransactions.Count,
                             tx.Id,
-                            txAddedBlockEncoding.EncodingLength,
+                            txAddedEncoding.EncodingLength,
                             maxBlockBytes);
                         continue;
                     }
-
-                    if (toMineCounts[tx.Signer] >= maxTransactionsPerSigner)
+                    else if (toMineCounts[tx.Signer] >= maxTransactionsPerSigner)
                     {
                         _logger.Debug(
                             "Ignoring tx {Iter}/{Total} {TxId} due to the maximum number " +
@@ -370,7 +341,7 @@ namespace Libplanet.Blockchain
                     transactionsToMine.Add(tx);
                     nextNonces[tx.Signer] += 1;
                     toMineCounts[tx.Signer] += 1;
-                    estimatedEncoding = txAddedBlockEncoding;
+                    estimatedEncoding = txAddedEncoding;
                 }
                 else if (tx.Nonce < storedNonces[tx.Signer])
                 {