+++ Libplanet/Blockchain/Policies/IBlockPolicy.cs
@@ -103,11 +103,12 @@ namespace Libplanet.Blockchain.Policies
         long GetNextBlockDifficulty(BlockChain<T> blockChain);
 
         /// <summary>
-        /// Gets the maximum length of a <see cref="Block{T}"/> in bytes.
+        /// Gets the maximum length of <see cref="Block{T}.Transactions"/> in bytes.
         /// </summary>
         /// <param name="index">The <see cref="Block{T}.Index"/> of the <see cref="Block{T}"/>
         /// for which this constraint should apply.</param>
-        /// <returns>The maximum length of a <see cref="Block{T}"/> in bytes to accept.</returns>
+        /// <returns>The maximum length of <see cref="Block{T}.Transactions"/> in bytes
+        /// to accept.</returns>
         long GetMaxBlockBytes(long index);
 
         /// <summary>